requirements.txt
# Deep Learning & ML Framework
torch>=2.0.0
torchvision>=0.15.0

# Web UI & Visualization
streamlit>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing & ML Utilities
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0